requirements.txt
# ─── Core ────────────────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ─── DSPy Agent Framework ─────────────────────────────────────
dspy-ai>=2.4.0
openai>=1.30.0

# ─── Web Search ───────────────────────────────────────────────
duckduckgo-search>=6.1.0
requests>=2.31.0

# ─── Document Processing ──────────────────────────────────────
pypdf>=4.2.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# PDF to images (optional, requires poppler system dependency)
pdf2image>=1.17.0
Pillow>=10.3.0

# PyMuPDF as fallback for pdf2image
PyMuPDF>=1.24.0

# ─── Embeddings + Vector Search ───────────────────────────────
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# ─── Topic Modeling ───────────────────────────────────────────
bertopic>=0.16.0
scikit-learn>=1.4.0

# ─── Data Processing ──────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0

# ─── Excel Generation ─────────────────────────────────────────
openpyxl>=3.1.0

# ─── Misc ─────────────────────────────────────────────────────
tqdm>=4.66.0
